==> 01-VERIFY-Create_Initial_DB.ipynb <==
# Inisial Data

import requests
from bs4 import BeautifulSoup
import pandas as pd
import re
import string

def cleaning_content(content):
  content = re.sub(r'[^\x00-\x7F]+', ' ', content)
  content = re.sub(r'@\w+', '', content)
  content = content.lower()
  content = re.sub(r'[%s]' % re.escape(string.punctuation), ' ', content)

  return content

def convert_datetime(date):
  date = date.split()
  
  if date[0] == 'January':
    date[0] = '01'
  elif date[0] == 'February':
    date[0] = '02'
  elif date[0] == 'March':
    date[0] = '03'
  elif date[0] == 'April':
    date[0] = '04'
  elif date[0] == 'May':
    date[0] = '05'
  elif date[0] == 'June':
    date[0] = '06'
  elif date[0] == 'July':
    date[0] = '07'
  elif date[0] == 'August':
    date[0] = '08'
  elif date[0] == 'September':
    date[0] = '09'
  elif date[0] == 'October':
    date[0] = '10'
  elif date[0] == 'November':
    date[0] = '11'
  elif date[0] == 'December':
    date[0] = '12'

  tgl = date[1].replace(',', '')

  if tgl == '1':
    tgl= '01'
  if tgl == '2':
    tgl = '02'
  if tgl == '3':
    tgl = '03'
  if tgl == '4':
    tgl = '04'
  if tgl == '5':
    tgl = '05'
  if tgl == '6':
    tgl = '06'
  if tgl == '7':
    tgl = '07'
  if tgl == '8':
    tgl = '08'
  if tgl == '9':
    tgl = '09'

  new_date = '{}/{}/{}'.format(date[2], date[0], tgl)

  return new_date

judul = []
date = []
url = []
content_full = []
content_clear = []

page = 358
while page > 1:
  link = 'https://turnbackhoax.id/page/{}/'.format(page)
  
  session = requests.Session()
  res = session.get(link)
  soup = BeautifulSoup(res.content)

  news_ = soup.find_all('article', class_='mh-loop-item')

  for news in news_:
    h = news.find('a', {'rel': 'bookmark'})
    j = h.get_text().strip()
    d = news.find('span', class_='mh-meta-date updated').text.strip()
    d = convert_datetime(d)
    u = h['href']

    content_url = requests.get(u)
    content_soup = BeautifulSoup(content_url.content)
    content_list = []
    for i in content_soup.find('div', {'class':'entry-content mh-clearfix'}).find_all('p'):
        content_list.append(i.text)
    content = ''
    c_f = content.join(content_list)
    c_c = cleaning_content(content.join(content_list))

    print(j)
    judul.append(j)
    print(d)
    date.append(d)
    print(u)
    url.append(u)
    print('')
    content_full.append(c_f)
    content_clear.append(c_c)
  
  page -= 1

db = pd.DataFrame(judul, columns=['Judul'])
db['Tanggal'] = date
db['Link'] = url
db['Isi berita'] = content_full
db['Isi berita bersih'] = content_clear

db['Tanggal'] = pd.to_datetime(db['Tanggal'])

db = db.sort_values('Tanggal')
db

db.to_csv('dB_tbh.csv', index=False)
