==> 02-VERIFY-Data_Cleansing.ipynb <==
# Data Cleansing

Melakukan data cleansing terhadap verify_dataset yang terdiri atas:

1. Case folding.
2. Menghilangkan kata-kata di dalam kurung.
3. Menghilangkan _punctuation_.
4. Menghilangkan stopwords.
5. Melakukan _stemming_.

Melakukan import library yang dibutuhkan.

!pip install Sastrawi

import pandas as pd
import re
import string
from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

df = pd.read_csv('verify_dataset.csv')

df.sample(5)

## Text Cleaning

Text cleaning terdiri atas data cleansing 1, 2, dan 3. Ketiga tahap tersebut dipilih berdasarkan asesmen yang dilakukan saat melakukan eksplorasi data.

Membuat fungsi text_cleaning.

def text_cleaning(judul):
  judul = judul.lower()                                               # casefolding
  judul = re.sub('\(.*?\) | \[.*?\]', '', judul)                      # kata di dalam kurung
  judul = re.sub('[%s]' % re.escape(string.punctuation), '', judul)   # punctuation
  judul = re.sub('[‘’“”…]', '', judul)                                # tanda kutip

  return judul

Mengaplikasikan fungsi text_cleaning ke dalam dataset.

df['Judul'] = df['Judul'].apply(text_cleaning)

df.sample(5)

## Stopwords

Tahap ini berarti menghilangkan kata-kata yang sering muncul dan tidak terlalu berpengaruh kepada makna kalimat. Tahap ini menggunakan library sastrawi.

Load stopword dari sastrawi.

factory = StopWordRemoverFactory()
new_stopwords = ['yg', 'dgn', 'dlm', 'nya']
stopwords = factory.get_stop_words() + new_stopwords
stopwords = factory.create_stop_word_remover()

#for stopword in stopwords:
#  print(stopword)

Membuat fungsi stopwords_removal.

def stopwords_removal(judul):
  judul = stopwords.remove(judul)

  return judul

Mengaplikasikan fungsi stopwords_removal kepada dataset.

df['Judul'] = df['Judul'].apply(stopwords_removal)

df.sample(5)

## Steming

Tahap steming merupakan penghapusan imbuhan dalam suatu kata. Tahap ini juga menggunakan library sastrawi.

factory = StemmerFactory()
stemmer = factory.create_stemmer()

Mengaplikasikan stemmer ke dalam dataset.

df['Judul'] = df['Judul'].apply(lambda judul:stemmer.stem(judul))

df.sample(5)

## Save verify_dataset_clean

Menyimpan dataset yang telah melalui tahap text_cleansing.

df.to_csv('verify_dataset_clean.csv', index=False)
