==> 03-VERIFY-Tuning_Hyperparameter.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.pipeline import Pipeline

# TF IDF

df = pd.read_csv('verify_dataset_clean.csv')
df.sample(5)

x = df['Judul']
y = df['Label']

tvec_pipe = Pipeline([
                  ('tfidf', TfidfVectorizer()),
                  ('logreg', LogisticRegression())
])
tvec_params = {
    'tfidf__ngram_range' : ((1,1), (1,2), (2,2)),
    'tfidf__max_df' : (0.5, 0.75, 1.0),
    'tfidf__max_features' : (None, 5000, 10000, 20000),
    'tfidf__norm' : ('l1', 'l2'),
    'tfidf__use_idf' : (True, False)
}

tvec_gs = GridSearchCV(tvec_pipe, tvec_params, cv = 5, verbose = 1, n_jobs = -1)

tvec_gs.fit(x, y)

print(tvec_gs.best_params_)

tvec_gs.best_score_

## Mengaplikasikan hyperparameter TF-IDF terbaik

tfidf = TfidfVectorizer(
    ngram_range = (1,2),
    max_df = 0.5,
    use_idf = True,
    max_features = None,
    norm = 'l2'
    )

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.1)

x_train_tfidf = tfidf.fit_transform(x_train)
x_test_tfidf = tfidf.transform(x_test)

print(len(x_test))
x_test_tfidf.shape

# Logistic Regression

model = LogisticRegression()
model_params = {
    'penalty' : ['l1', 'l2'],
    'C' : np.logspace(-1,4,10),
    'solver' : ['liblinear', 'saga', 'sag', 'lbfgs'],
    'max_iter' : [5000, 10000]
}

logreg_gs = GridSearchCV(model, model_params, cv = 5, verbose = 1, n_jobs = -1)

logreg_gs.fit(x_train_tfidf, y_train)

print(logreg_gs.best_params_)

logreg_gs.best_score_

# SVC
