==> 04-VERIFY-Build_Logistic_Regression.ipynb <==
import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_score, StratifiedKFold, train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer, TfidfTransformer
from matplotlib import pyplot as plt
import seaborn as sns
import pickle
import numpy as np
from sklearn.metrics import accuracy_score, f1_score, precision_score

df = pd.read_csv('verify_dataset_clean.csv')
df.sample(5)

x_train_text, x_test_text, y_train, y_test = train_test_split(df['Judul'], df['Label'], test_size=0.1)

vectorizer = CountVectorizer(
    ngram_range = (1, 2),
    decode_error = 'replace',
    max_df = 0.5,
    max_features = None
)

pickle.dump(vectorizer.vocabulary, open('fitur.pkl', 'wb'))

transformer = TfidfTransformer(
    norm = 'l2',
    use_idf = True
)

#loaded_vec = CountVectorizer(decode_error="replace",vocabulary=pickle.load(open("fitur.pkl", "rb")))

x_train = transformer.fit_transform(vectorizer.fit_transform(np.array(x_train_text)))
x_test = transformer.transform(vectorizer.transform(np.array(x_test_text)))

x_train.shape, x_test.shape

len(y_train), len(y_test)

verify_logreg = LogisticRegression(
    C = 774.2636826811278, 
    solver='saga', 
    penalty = 'l1',
    max_iter = 5000
    )

skf = StratifiedKFold(
    n_splits=5,
    shuffle=True,
    random_state=17
    )

cv_results = cross_val_score(
    verify_logreg, 
    x_train, 
    y_train, 
    cv=skf, 
    scoring='accuracy'
    )

cv_results, cv_results.mean()

verify_logreg.fit(x_train, y_train)

test_preds = verify_logreg.predict(x_test)

dfP = pd.DataFrame(x_test_text, columns=['Judul'])
dfP['Predict'] = test_preds
dfP['True'] = y_test

dfP.head()

def my_metrics(y_true, y_pred):
    accuracy=accuracy_score(y_true, y_pred)
    precision=precision_score(y_true, y_pred,average='weighted')
    f1Score=f1_score(y_true, y_pred, average='weighted') 
    print("Accuracy  : {}".format(accuracy))
    print("Precision : {}".format(precision))
    print("f1Score : {}".format(f1Score))
    
    return accuracy, precision, f1Score

my_metrics(y_test, test_preds)

confusion_matrix = pd.crosstab(dfP['True'], dfP['Predict'], rownames=['Actual'], colnames=['Predicted'])
print(confusion_matrix)
#sns.heatmap(confusion_matrix, annot=True)
#plt.show()

verify_logreg_model = 'verify_logreg_model.sav'
pickle.dump(verify_logreg, open(verify_logreg_model, 'wb'))

predict_proba = verify_logreg.predict_proba(x_test)

i = 0
hoax_prob_long = []
for hoax in predict_proba:
  hoax_prob_long.append(predict_proba[i,1]*100)
  i+=1

hoax_prob = []
for j in hoax_prob_long:
  hoax_prob.append('{:.1f}'.format(j))

dfP['Hoax Probaility'] = hoax_prob

dfP.head()

dfP.loc[dfP['True'] == 1, 
       'True'] = 'HOAX'
dfP.loc[dfP['True'] == 0, 
       'True'] = 'BUKAN HOAX'

dfP_final = dfP.drop(columns=['Predict'])

dfP_final.columns = ['Judul', 'Actual', 'Hoax Probability']

dfP_final = dfP_final[['Judul', 'Hoax Probability', 'Actual']]

dfP_final.head()

dfP_final.to_csv('hasil_verify_logreg_model.csv', index=False)
