==> 05-VERIFY-Predict_Logistic_Regression.ipynb <==
!pip install Sastrawi

import pickle
import re
import string
from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory
from sklearn.feature_extraction.text import TfidfVectorizer

def preprocessing(judul):
  judul = judul.lower()                                               # casefolding
  judul = re.sub('\(.*?\) | \[.*?\]', '', judul)                      # kata di dalam kurung
  judul = re.sub('[%s]' % re.escape(string.punctuation), '', judul)   # punctuation
  judul = re.sub('[‘’“”…]', '', judul)                                # tanda kutip

  factory = StopWordRemoverFactory()
  new_stopwords = ['yg', 'dgn', 'dlm', 'nya']
  stopwords = factory.get_stop_words() + new_stopwords
  stopwords = factory.create_stop_word_remover()
  judul = stopwords.remove(judul)

  factory = StemmerFactory()
  stemmer = factory.create_stemmer()
  judul = stemmer.stem(judul)

  return judul

def ekstraksi_fitur(judul):
  transformer = pickle.load(open("transformer.pickle", "rb"))
  judul = transformer.transform([judul])

  return judul

def verify_predict_logreg(judul):
  
  judul = preprocessing(judul)
  fitur = ekstraksi_fitur(judul)
  model = pickle.load(open('verify_logreg_model.sav', 'rb'))

  hoax_prob = model.predict_proba(fitur)[i,1]*100

  return hoax_prob

to_predict = 'Anies Baswedan Donasikan Dana untuk Palestina dari Uang Pribadi dan Karyawan Pemda DKI'

judul = preprocessing(to_predict)

judul

transformer = pickle.load(open("transformer.pickle", "rb"))

test = transformer.fit_transform([judul])

test.shape

model = pickle.load(open('verify_logreg_model.sav', 'rb'))

pred = model.predict(test)
