==> 09-VERIFY-Predict_tf.ipynb <==
import pickle
from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer
import tensorflow as tf
import numpy as np

vectorizer = CountVectorizer(vocabulary=pickle.load(open("fitur.pkl", "rb")))
vectorizer._validate_vocabulary()
transformer = TfidfTransformer(
    norm = 'l2',
    use_idf = True
)

user_input = 'Jokowi berniat mengundurkan diri dari jabatan presiden bulan depan'
x = transformer.transform(vectorizer.transform(np.array([user_input]))).toarray()

model = tf.keras.models.load_model('verify_model.h5', compile=False)

prediksi_hoax = model.predict(x)
prediksi_hoax *= 100
prediksi_hoax
